40-gai-stack/langchain/pyproject.toml
[project]
name = "langchain_motex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.17.0",
    "flashrank>=0.2.10",
    "langchain>=0.3.17",
    "langchain-anthropic>=0.3.5",
    "langchain-chroma>=0.2.1",
    "langchain-community>=0.3.16",
    "langchain-core>=0.3.33",
    "langchain-elasticsearch>=0.3.2",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.3",
    "langchain-text-splitters>=0.3.5",
    "langchain-unstructured>=0.1.6",
    "langserve>=0.3.1",
    "markitdown>=0.0.1a3",
    "ollama>=0.4.7",
    "pdfplumber>=0.11.5",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "unstructured>=0.16.17",
    "unstructured-client>=0.29.0",
]

[tool.setuptools]
package-dir = {"" = "src/langchain_motex"}

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests",]
